Move URILiteralUtils to compiler.utils package
