+++ src/main/java/org/rumbledb/compiler/utils/URILiteralUtils.java
@@ -5,7 +5,7 @@
  * The ASF licenses this file to You under the Apache License, Version 2.0.
  */
 
-package org.rumbledb.compiler;
+package org.rumbledb.compiler.utils;
 
 import org.rumbledb.exceptions.CannotRetrieveResourceException;
 import org.rumbledb.exceptions.ExceptionMetadata;
@@ -14,12 +14,12 @@ import org.rumbledb.runtime.functions.input.FileSystemUtil;
 
 import java.net.URI;
 
-final class URILiteralUtils {
+public final class URILiteralUtils {
 
     private URILiteralUtils() {
     }
 
-    static URI resolve(URI baseURI, String literal, ExceptionMetadata metadata) {
+    public static URI resolve(URI baseURI, String literal, ExceptionMetadata metadata) {
         try {
             return FileSystemUtil.resolveURI(baseURI, literal, metadata);
         } catch (CannotRetrieveResourceException exception) {
